normalise trailing dot
This is version 2 of a patch for ticket #3975

This patch modifles normalise to treat paths with a trailing dot as the
corresponding directory: the trailing . is stripped, and the pathSeparator
is appended to the result of joining path components.

The special case of all path components being "." is handled first, replacing
the existing special case test for paths with a leading dot. Dots are now
dropped consistently for all path components.

Additional tests are provided for:

    normalise "./." == "./"
    normalise "bob/fred/." == "bob/fred/"