Refactor normalise: simplify propSep
Only when a path on Windows starts with a leading forward slash does propSep
need to do something, all others path separators are properly taken care of by
joinPath.

Add 2 tests to show that splitDirectories can handle superfluous
pathSeparators, and a test to show that (this new version of) propSep is
indeed needed to normalise relative-to-root paths on Windows.