Revert "Avoid partial functions Data.List.init and Data.List.last"
This reverts commit 19623874284b7de90a1e37371521d41c486af840.

I should have run benchmarks when working on the original commit:
now `hasTrailingPathSeparator` regressed a lot, because
`getTrailingPathSeparator` does not fuse well and we are left
with a lot of overhead. And if we revert `hasTrailingPathSeparator` back
and plug `-Wno-x-partial` to accomodate future warnings on `Data.List.{init,last}`,
then there is no raison d'etre for other changes, because
we will have `-Wno-x-partial` anyway.

In general this performance issue could be worked out / worked around,
but we have a very limited toolbox because of keeping both `FilePath`
and `OsPath` implementations in the same place.

So for now let's revert (and add {-# OPTIONS_GHC -Wno-x-partial #-}
once `Data.List.{init,last}` gain warnings).

Additional context: https://github.com/haskell/core-libraries-committee/issues/292